requirements.txt
boto3
openai-whisper
numpy
torch
psycopg2-binary